Move tags_with_prefix() to a helpers.py module
