+++ src/settlements.py
@@ -2,6 +2,7 @@ import sys
 import pyarrow
 
 from .geoparquet import GeoParquetWriter
+from .helpers import tags_with_prefix
 
 
 class SettlementsWriter(GeoParquetWriter):
@@ -73,12 +74,3 @@ def column(column_name, tags):
             return tags_with_prefix(f"{column_name[:-1]}:", tags)
         case _:
             return tags.get(column_name)
-
-
-def tags_with_prefix(prefix, tags):
-    """
-    Returns a dict of all tags with the given prefix string; keys in the
-    dict will have the prefix dropped.
-    """
-    prefix_len = len(prefix)
-    return {k[prefix_len:]: v for (k, v) in tags if k.startswith(prefix)}
